Set up initial 1-Player Game
I created the initial 1-player game of snake with wasd controls for
movement. This updates every clock tick based on the snake timer, we can
extend this to accept ai input after the player input in the if/else
statements with arrow key input.

The tutorial I followed also included how to print messages and accept
menu input, so we may not need to use TKinter for the interface aspect.