Updated game to incorporate 2 player snake and updated size
I updated the game with a second player for the snake and broke out some
functionality into functions to be reusable. The second player uses
arrow keys but this input will later just be direct state changes by the
ai.

I also changed the size of the board to 600x600 with block size 20x20
to maintain the 30x30 grid needed for ai gym to work correctly.